Make `Forest` and related types completely polymorphic
Note that the `Forest` with different ID types with the same internal ID
types will be monomorphized separately. This means that the codes would
be monomorphized in redundant way, since the logics for `Forest` mostly
depends on `Id::Internal`, but not `Id`. `Id` will only affects to the
returned types and formatting.

However, this issue is implementation detail and the crate itself should
work perfectly fine with the current implementation, so I consider this
is worth merging to develop branch and releasing.

I'll improve this once I have great idea to reduce the redundant
monomorphization while keeping maintainability. Such change should be
non-breaking, and can be safely delayed.